remove unnecessary module_utils, move the classes into the module code
